move query plan execution tests to a separate file
